requirements.txt
# Python dependencies for PumpIQ
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
asyncpg==0.29.0

# AI/ML
google-genai>=1.0.0
openai==1.10.0
tiktoken==0.5.2

# Data Processing
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2

# API Clients
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Task Queue & Scheduling
celery==5.3.6
celery[redis]==5.3.6
apscheduler==3.10.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Validation & Parsing
jsonschema==4.20.0
email-validator==2.1.0
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # for testing FastAPI
faker==22.5.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
pytz==2024.1
python-slugify==8.0.1
cachetools==5.3.2
